refactor: convert useAccount hook to context
useAccount hook meant logic was being ran in every component it is rendered in,
which resulted in many rpc calls when setting up a privy account
