+++ packages/interface/src/components/SingleRoundHome.tsx
@@ -5,8 +5,8 @@ import { Info } from "~/components/Info";
 import { JoinButton } from "~/components/JoinButton";
 import { Button } from "~/components/ui/Button";
 import { Heading } from "~/components/ui/Heading";
+import { useAccount } from "~/contexts/Account";
 import { useMaci } from "~/contexts/Maci";
-import useAccount from "~/hooks/useAccount";
 import { useIsMobile } from "~/hooks/useIsMobile";
 import { useRoundState } from "~/utils/state";
 import { ERoundState, type IRoundData } from "~/utils/types";
