+++ packages/interface/src/contexts/Ballot.tsx
@@ -1,10 +1,9 @@
 import React, { createContext, useContext, useState, useEffect, useMemo, useCallback } from "react";
 
-import useAccount from "~/hooks/useAccount";
-
 import type { BallotContextType, BallotProviderProps } from "./types";
 import type { Ballot, Vote } from "~/features/ballot/types";
 
+import { useAccount } from "./Account";
 import { useRound } from "./Round";
 
 export const BallotContext = createContext<BallotContextType | undefined>(undefined);
