+++ packages/interface/src/features/ballot/components/BallotConfirmation.tsx
@@ -11,10 +11,10 @@ import { Button } from "~/components/ui/Button";
 import { Heading } from "~/components/ui/Heading";
 import { Notice } from "~/components/ui/Notice";
 import { config } from "~/config";
+import { useAccount } from "~/contexts/Account";
 import { useBallot } from "~/contexts/Ballot";
 import { useRound } from "~/contexts/Round";
 import { useProjectCount } from "~/features/projects/hooks/useProjects";
-import useAccount from "~/hooks/useAccount";
 import { formatNumber } from "~/utils/formatNumber";
 import { useRoundState } from "~/utils/state";
 import { ERoundState } from "~/utils/types";
