+++ packages/interface/src/features/proposals/components/MetadataForm.tsx
@@ -5,7 +5,7 @@ import { toast } from "sonner";
 import { ImageUpload } from "~/components/ImageUpload";
 import { FieldArray, Form, FormControl, FormSection, Select, Textarea } from "~/components/ui/Form";
 import { Input } from "~/components/ui/Input";
-import useAccount from "~/hooks/useAccount";
+import { useAccount } from "~/contexts/Account";
 import { useIsCorrectNetwork } from "~/hooks/useIsCorrectNetwork";
 
 import { useCreateProposal } from "../hooks/useCreateProposal";
