+++ packages/interface/src/features/rounds/components/Projects.tsx
@@ -9,10 +9,10 @@ import { SortFilter } from "~/components/SortFilter";
 import { StatusBar } from "~/components/StatusBar";
 import { Button } from "~/components/ui/Button";
 import { Heading } from "~/components/ui/Heading";
+import { useAccount } from "~/contexts/Account";
 import { useBallot } from "~/contexts/Ballot";
 import { useMaci } from "~/contexts/Maci";
 import { useRound } from "~/contexts/Round";
-import useAccount from "~/hooks/useAccount";
 import { useMyProjects } from "~/hooks/useProjects";
 import { useResults } from "~/hooks/useResults";
 import { useRoundState } from "~/utils/state";
