+++ packages/interface/src/hooks/useIsCorrectNetwork.ts
@@ -1,8 +1,7 @@
 import { useChainId } from "wagmi";
 
 import { config } from "~/config";
-
-import useAccount from "./useAccount";
+import { useAccount } from "~/contexts/Account";
 
 export interface IUseIsCorrectNetworkReturn {
   isCorrectNetwork: boolean;
