+++ packages/interface/src/layouts/DefaultLayout.tsx
@@ -5,10 +5,10 @@ import Header from "~/components/Header";
 import { Info } from "~/components/Info";
 import { Notice } from "~/components/ui/Notice";
 import { config } from "~/config";
+import { useAccount } from "~/contexts/Account";
 import { useBallot } from "~/contexts/Ballot";
 import { useMaci } from "~/contexts/Maci";
 import { SubmitBallotButton } from "~/features/ballot/components/SubmitBallotButton";
-import useAccount from "~/hooks/useAccount";
 import { useRoundState } from "~/utils/state";
 import { ERoundState } from "~/utils/types";
